refactor: rename BasicLocationPuck to LocationPuck and provide locationPuckColors extension for material3 instead
